+++ app/(afterLogin)/messages/[userId]/_components/UserInfo.tsx
